requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-ckeditor==6.4.2
django-js-asset==2.0.0
pkg_resources==0.0.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
